requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.4
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==1.0.0
dj-rest-auth==2.2.4
Django==4.0.6
django-allauth==0.51.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-heroku==0.3.1
djangorestframework==3.13.1
environ==1.0
et-xmlfile==1.1.0
gcloud==0.18.3
googleapis-common-protos==1.56.4
gunicorn==20.1.0
h11==0.13.0
httplib2==0.20.4
idna==3.3
jwcrypto==1.4.2
oauth2client==4.1.3
oauthlib==3.2.0
openpyxl==3.0.10
Pillow==9.2.0
protobuf==4.21.6
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
PyJWT==2.4.0
pyparsing==3.0.9
Pyrebase4==4.5.0
python-dotenv==0.20.0
python-jwt==3.3.4
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.10
uvicorn==0.18.2
whitenoise==6.2.0
wrapt==1.14.1
